rename unique to filterOutFalsy
