+++ src/logic/findRemovedFieldAndRemoveListener.ts
@@ -7,7 +7,7 @@ import isCheckBoxInput from '../utils/isCheckBoxInput';
 import isDetached from '../utils/isDetached';
 import isArray from '../utils/isArray';
 import unset from '../utils/unset';
-import unique from '../utils/unique';
+import filterOutFalsy from '../utils/filterOutFalsy';
 import isUndefined from '../utils/isUndefined';
 import { Field, FieldRefs, FieldValues, Ref } from '../types/form';
 
@@ -48,7 +48,7 @@ export default function findRemovedFieldAndRemoveListener<
     const { options } = fieldRef;
 
     if (isArray(options) && options.length) {
-      unique(options).forEach((option, index): void => {
+      filterOutFalsy(options).forEach((option, index): void => {
         const { ref, mutationWatcher } = option;
         if ((ref && isDetached(ref) && isSameRef(option, ref)) || forceDelete) {
           removeAllEventListeners(ref, handleChange);
@@ -61,7 +61,7 @@ export default function findRemovedFieldAndRemoveListener<
         }
       });
 
-      if (options && !unique(options).length) {
+      if (options && !filterOutFalsy(options).length) {
         delete fieldsRef.current[name];
       }
     } else {
